+++ service/controller/resource/tcnp/resource.go
@@ -1,13 +1,16 @@
 package tcnp
 
 import (
+	"encoding/json"
+
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 
+	"github.com/giantswarm/aws-operator/service/controller/resource/tcnp/template"
 	"github.com/giantswarm/aws-operator/service/internal/changedetection"
 	"github.com/giantswarm/aws-operator/service/internal/images"
-	event "github.com/giantswarm/aws-operator/service/internal/recorder"
+	"github.com/giantswarm/aws-operator/service/internal/recorder"
 )
 
 const (
@@ -17,11 +20,12 @@ const (
 
 type Config struct {
 	Detection *changedetection.TCNP
-	Event     event.Interface
+	Event     recorder.Interface
 	Images    images.Interface
 	K8sClient k8sclient.Interface
 	Logger    micrologger.Logger
 
+	AlikeInstances   string
 	InstallationName string
 }
 
@@ -29,11 +33,12 @@ type Config struct {
 // Plane. We manage a dedicated Cloud Formation stack for each node pool.
 type Resource struct {
 	detection *changedetection.TCNP
-	event     event.Interface
+	event     recorder.Interface
 	images    images.Interface
 	k8sClient k8sclient.Interface
 	logger    micrologger.Logger
 
+	alikeInstances   map[string][]template.LaunchTemplateOverride
 	installationName string
 }
 
@@ -54,8 +59,19 @@ func New(config Config) (*Resource, error) {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
 
+	if config.AlikeInstances == "" {
+		return nil, microerror.Maskf(invalidConfigError, "%T.AlikeInstances must not be empty", config)
+	}
 	if config.InstallationName == "" {
 		return nil, microerror.Maskf(invalidConfigError, "%T.InstallationName must not be empty", config)
+	}
+
+	var alikeInstances map[string][]template.LaunchTemplateOverride
+	{
+		err := json.Unmarshal([]byte(config.AlikeInstances), &alikeInstances)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
 	}
 
 	r := &Resource{
@@ -65,6 +81,7 @@ func New(config Config) (*Resource, error) {
 		k8sClient: config.K8sClient,
 		logger:    config.Logger,
 
+		alikeInstances:   alikeInstances,
 		installationName: config.InstallationName,
 	}