+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Make it mandatory to configure alike instances via e.g. the installations repo.
+
 ## [9.3.1] - 2020-11-12
 
 ### Changed